Split CSS code from component file
