+++ src/components/Action/ChangeAddressAction.js
@@ -1,4 +1,5 @@
 import Action from '../Action.js';
+import changeAddressActionStyle from '../../styles/components/action/changeAddressAction.js';
 
 class ChangeAdressAction extends Action {
   constructor(id) {
@@ -7,28 +8,12 @@ class ChangeAdressAction extends Action {
     const shadow = this.shadowRoot;
 
     shadow.innerHTML = `
-    <action-element id="${id}" mainBtn="주소 변경" subBtn="계속" pointerPosition="left">
-      <span slot="text">KR으로 배송할 품목을 표시하겠습니다. 다른 국가로 배송되는 품목을 보려면 배송 주소를 변경하십시오.</span>
-    </action-element>
+      <action-element id="${id}" mainBtn="주소 변경" subBtn="계속" pointerPosition="left">
+        <span slot="text">KR으로 배송할 품목을 표시하겠습니다. 다른 국가로 배송되는 품목을 보려면 배송 주소를 변경하십시오.</span>
+      </action-element>
     `;
 
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-    const type = this.getAttribute('type');
-
-    style.textContent = `
-      :host {
-        color: var(--black);
-      }
-
-      action-element {
-        z-index: 1;
-      }
-    `;
-    return style;
+    this.shadowRoot.append(changeAddressActionStyle());
   }
 }
