+++ src/components/BackDrop.js
@@ -1,3 +1,5 @@
+import backdropStyle from '../styles/components/backdrop.js';
+
 class Backdrop extends HTMLElement {
   constructor() {
     super();
@@ -5,24 +7,7 @@ class Backdrop extends HTMLElement {
     this.attachShadow({ mode: 'open' });
     this.isOpen = false;
 
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-
-    style.textContent = `
-      :host {
-        width: 100%;
-        height: 100%;
-        display: block;
-        background-color: var(--black-40);
-        position: fixed;
-        top: 0;
-        left: 0;
-      }
-    `;
-    return style;
+    this.shadowRoot.append(backdropStyle());
   }
 }
