+++ src/components/NavBar/SubMenu.js
@@ -1,3 +1,5 @@
+import subMenuStyle from '../../styles/components/navbar/subMenu.js';
+
 class SubMenu extends HTMLElement {
   constructor() {
     super();
@@ -17,53 +19,7 @@ class SubMenu extends HTMLElement {
       </button>
     `;
 
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-
-    style.textContent = `
-      :host {
-        display: flex;
-        align-items: center;
-        color: var(--white);
-      }
-
-      button {
-        background-color: transparent;
-        border: none;
-
-        display: flex;
-        padding: 8px;
-        gap: 4px;
-        align-items: center;
-
-        color: var(--white);
-        text-decoration: none;
-        
-        font-weight: 600;
-        font-size: var(--bold-md-size);
-        line-height: 16px;
-        letter-spacing: var(--bold-md-spacing);
-
-        border-radius: 1px;
-      }
-
-      button:hover {
-        cursor: pointer;
-        outline: 1px solid var(--white);
-      }
-
-      span {
-        width: max-content;
-      }
-
-      .bold {
-        font-weight: 800;
-      }
-    `;
-    return style;
+    this.shadowRoot.append(subMenuStyle());
   }
 }
