--- src/components/Sidebar/SidebarList.js
@@ -1,23 +1,0 @@
-class SidebarList extends HTMLElement {
-  constructor() {
-    super();
-
-    const text = this.innerText;
-    const shadow = this.attachShadow({ mode: 'open' });
-
-    shadow.innerHTML = `<button>list</button>`;
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-    const type = this.getAttribute('type');
-
-    style.textContent = `
-     
-    `;
-    return style;
-  }
-}
-
-export default SidebarList;
