+++ src/components/Sidebar/SidebarMain.js
@@ -1,5 +1,5 @@
-import { SIDEBAR_SIZE } from '../../constant.js';
 import { menus } from '../../data/menu.js';
+import sidebarMainStyle from '../../styles/components/sidebar/sidebarMain.js';
 
 class SidebarMain extends HTMLElement {
   constructor() {
@@ -32,30 +32,30 @@ class SidebarMain extends HTMLElement {
             menu.categories.length
           );
           return `
-              <section class="section" id="${menu.id}">
-                <sidebar-title-element>${menu.title}</sidebar-title-element>
-                ${menuDefault
-                  .map((category) => {
-                    return `<sidebar-category-element id=${category.id}>${category.name}</sidebar-category-element>`;
-                  })
-                  .join('')}
-                <div id="folding-container">
-                  <sidebar-fold-element id="unfolidng-btn"></sidebar-fold-element>
-                  <ul id="folidng-list">
-                    ${menuOverflowed
-                      .map((category) => {
-                        return `<sidebar-category-element id=${category.id}>${category.name}</sidebar-category-element>`;
-                      })
-                      .join('')}
-                    <sidebar-unfold-element id="folidng-btn"></sidebar-unfold-element>
-                  </ul>
-                </div>
-              </section>`;
+            <section class="section" id="${menu.id}">
+              <sidebar-title-element>${menu.title}</sidebar-title-element>
+              ${menuDefault
+                .map((category) => {
+                  return `<sidebar-category-element id=${category.id}>${category.name}</sidebar-category-element>`;
+                })
+                .join('')}
+              <div id="folding-container">
+                <sidebar-fold-element id="unfolidng-btn"></sidebar-fold-element>
+                <ul id="folidng-list">
+                  ${menuOverflowed
+                    .map((category) => {
+                      return `<sidebar-category-element id=${category.id}>${category.name}</sidebar-category-element>`;
+                    })
+                    .join('')}
+                  <sidebar-unfold-element id="folidng-btn"></sidebar-unfold-element>
+                </ul>
+              </div>
+            </section>`;
         })
         .join('')}
     `;
 
-    this.shadowRoot.append(this.getStyle());
+    this.shadowRoot.append(sidebarMainStyle());
   }
 
   createTitle(parent, text) {
@@ -68,76 +68,6 @@ class SidebarMain extends HTMLElement {
     const category = document.createElement('sidebar-category-element');
     category.shadowRoot.querySelector('p').innerHTML = text;
     parent.append(category);
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-    const type = this.getAttribute('type');
-
-    style.textContent = `
-      :host {
-        display: block;
-        width: ${SIDEBAR_SIZE}px;
-        height: 100%;
-      }
-
-      #folidng-list {
-        padding: 0;
-        list-style: none;
-        margin: 0;
-        opacity: 0;
-        height: 0;
-        overflow: hidden;
-      }
-
-      #folidng-list.unfolded {
-        opacity: 1;
-        height: auto;
-      }
-
-      #folidng-list.slide-down {
-        animation: slide-down .3s forwards;
-      }
-
-      #folidng-list.slide-up {
-        animation: slide-up .3s forwards;
-      }
-
-      ul {
-        margin: 0;
-        padding: 0;
-      }
-
-      .section {
-        border-bottom: 1px solid var(--gray-100);
-      }
-
-      .section:last-of-type {
-        border-bottom: none;
-      }
-
-      @keyframes slide-up {
-        0% {
-          opacity: 1;
-          height: auto;
-        }
-        100% {
-          opacity: 0;
-          height: 0;
-        }
-      }
-
-      @keyframes slide-down {
-        0% {
-          opacity: 0;
-          height: 0;
-        }
-        100% {
-          opacity: 1;
-          height: auto;
-        }
-    `;
-    return style;
   }
 }
