+++ src/components/Sidebar/SidebarMain/SidebarUnfold.js
@@ -1,0 +1,14 @@
+import sidebarUnfoldStyle from '../../../styles/components/sidebar/sidebarMain/sidebarUnfold.js';
+
+class SidebarUnfold extends HTMLElement {
+  constructor() {
+    super();
+
+    const shadow = this.attachShadow({ mode: 'open' });
+
+    shadow.innerHTML = `<button>간단히 보기<icon-element name="chevron-down"></icon-element></button>`;
+    this.shadowRoot.append(sidebarUnfoldStyle());
+  }
+}
+
+export default SidebarUnfold;
