+++ src/components/Sidebar/SidebarSub/SidebarSubContent.js
@@ -1,4 +1,4 @@
-import { SIDEBAR_SIZE } from '../../constant.js';
+import sidebarSubContentStyle from '../../../styles/components/sidebar/sidebarSub/sidebarSubContent.js';
 
 class SidebarSub extends HTMLElement {
   constructor(submenus) {
@@ -14,7 +14,6 @@ class SidebarSub extends HTMLElement {
           const categories = submenu.categories;
 
           return `
-          
             <section>
               <sidebar-title-element>${title}</sidebar-title-element>
               ${categories
@@ -30,34 +29,7 @@ class SidebarSub extends HTMLElement {
         .join('')}
     `;
 
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-    const type = this.getAttribute('type');
-
-    style.textContent = `
-      :host {
-        display: block;
-        width: ${SIDEBAR_SIZE}px;
-        height: 100%;
-      }
-
-      section {
-        border-top: 1px solid #e5e5e5;
-      }
-
-      @keyframes slide-right {
-        0% {
-          transform: translateX(100%);
-        }
-        100% {
-          transform: translateX(0%);
-        }
-      }
-    `;
-    return style;
+    this.shadowRoot.append(sidebarSubContentStyle());
   }
 }
