+++ src/components/Sidebar/SidebarTitle.js
@@ -1,3 +1,5 @@
+import sidebarTitleStyle from '../../styles/components/sidebar/sidebarTitle.js';
+
 class SideBarTitle extends HTMLElement {
   constructor() {
     super();
@@ -6,30 +8,7 @@ class SideBarTitle extends HTMLElement {
     const shadow = this.attachShadow({ mode: 'open' });
 
     shadow.innerHTML = `<p>${text}</p>`;
-    this.shadowRoot.append(this.getStyle());
-  }
-
-  getStyle() {
-    const style = document.createElement('style');
-    const type = this.getAttribute('type');
-
-    style.textContent = `
-      :host {
-        font-weight: var(--bold-lg-weight);
-        font-size: var(--bold-lg-size);
-        line-height: var(--bold-lg-height);
-        letter-spacing: var(--bold-lg-spacing);
-
-        display: flex;
-        align-items: center;
-        padding: 16px 32px;
-      }
-
-      p {
-        margin: 0;
-      }
-    `;
-    return style;
+    this.shadowRoot.append(sidebarTitleStyle());
   }
 }
