+++ src/components/index.js
@@ -12,16 +12,15 @@ import Action from './Action.js';
 import LoginAction from './Action/LoginAction.js';
 import ChangeAdressAction from './Action/ChangeAddressAction.js';
 import Backdrop from './Backdrop.js';
-import SideBarBack from './Sidebar/SidebarBack.js';
+import SideBarBack from './Sidebar/SidebarSub/SidebarBack.js';
 import SidebarCategory from './Sidebar/SidebarCategory.js';
-import SidebarFold from './Sidebar/SidebarFold.js';
-import SidebarUnfold from './Sidebar/SidebarUnFold.js';
+import SidebarFold from './Sidebar/SidebarMain/SidebarFold.js';
+import SidebarUnfold from './Sidebar/SidebarMain/SidebarUnfold.js';
 import SidebarHeader from './Sidebar/SidebarHeader.js';
-import SideBarList from './Sidebar/SidebarList.js';
 import SideBarTitle from './Sidebar/SidebarTitle.js';
 import SidebarMain from './Sidebar/SidebarMain.js';
 import SidebarSub from './Sidebar/SidebarSub.js';
-import SidebarSubContent from './Sidebar/SidebarSubContent.js';
+import SidebarSubContent from './Sidebar/SidebarSub/SidebarSubContent.js';
 
 customElements.define('btn-element', Button);
 customElements.define('footer-element', Footer);
@@ -41,7 +40,6 @@ customElements.define('sidebar-back-element', SideBarBack);
 customElements.define('sidebar-category-element', SidebarCategory);
 customElements.define('sidebar-fold-element', SidebarFold);
 customElements.define('sidebar-header-element', SidebarHeader);
-customElements.define('sidebar-list-element', SideBarList);
 customElements.define('sidebar-title-element', SideBarTitle);
 customElements.define('sidebar-main-element', SidebarMain);
 customElements.define('sidebar-sub-element', SidebarSub);
