+++ src/styles/components/action/changeAddressAction.js
@@ -1,0 +1,14 @@
+export default function changeAddressActionStyle() {
+  const style = document.createElement('style');
+
+  style.textContent = `
+      :host {
+        color: var(--black);
+      }
+
+      action-element {
+        z-index: 1;
+      }
+    `;
+  return style;
+}
